pyproject.toml
[tool.poetry]
name = "drf-homework"
version = "0.1.0"
description = ""
authors = ["Anton Gromov <seed14112@yaandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
generics = "^7.0.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
django-filter = "^24.3"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
